+++ src/Mail.php
@@ -120,7 +120,7 @@ class Mail extends AbstractPart
             foreach ($this->to as $address => $name) {
                 $to[] = $this->formatAddress($address, $name);
             }
-            $headers[] = 'To: ' . rtrim(implode(",\r\n ", $to));
+            $headers[] = $this->encodeHeader('To: ' . rtrim(implode(",\r\n ", $to)));
         }
 
         if (!empty($this->cc)) {
@@ -128,12 +128,12 @@ class Mail extends AbstractPart
             foreach ($this->cc as $address => $name) {
                 $cc[] = $this->formatAddress($address, $name);
             }
-            $headers[] = 'Cc: ' . rtrim(implode(",\r\n ", $cc));
+            $headers[] = $this->encodeHeader('Cc: ' . rtrim(implode(",\r\n ", $cc)));
         }
 
         if ($this->replyTo) {
             list($address, $name) = $this->replyTo;
-            $headers[] = 'Reply-To: ' . $this->formatAddress($address, $name);
+            $headers[] = $this->encodeHeader('Reply-To: ' . $this->formatAddress($address, $name));
         }
 
         if ($this->getPart() instanceof Mime\AbstractMime) {